Apply code review remarks
