+++ Server/package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "ejs": "^2.6.1",
     "express": "^4.16.4",
-    "lowdb": "^1.0.0",
     "models": "file:../Models",
     "protocol": "file:../Protocol",
     "utils": "file:../Utils"
