+++ Utils/src/logs.js
@@ -1,6 +1,6 @@
 const low = require('lowdb')
 const FileSync = require('lowdb/adapters/FileSync')
-const adapter = new FileSync('../logs.json')
+const adapter = new FileSync('logs.json')
 const db = low(adapter)
 
 function format(time) {
